pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "photo-ai-workflow"
version = "0.1.0"
description = "Three-stage photo culling workflow with technical checks, CLIP aesthetics, and optional VLM review"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26,<3",
  "pandas>=2.2,<3",
  "Pillow>=10.4,<12",
  "opencv-python>=4.10,<5",
  "torch>=2.6,<3",
  "transformers>=4.45,<5.0",
  "scikit-learn>=1.5,<2",
  "imagehash>=4.3,<5",
  "requests>=2.32,<3",
  "gradio>=4.44,<6",
  "python-dotenv>=1.0,<2"
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
photo-ai = "photo_ai_workflow.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
